requirements.txt
alpaca-py>=0.21.0
requests>=2.31.0
pandas>=2.0.0
numpy>=1.26.0
rich>=13.0.0
tenacity>=8.2.0
typer[all]>=0.12.3
python-dotenv>=1.0.1
pytest>=7.4.4
